dd6a5a8 comp: fix warning: cast to pointer from integer of different size
d521ffc comp: fix unused variable warnings

+++ src/ves/vesMapper.cpp
@@ -36,6 +36,7 @@
 #include <map>
 #include <vector>
 #include <cstdio>
+#include <stdint.h>
 
 class vesMapper::vesInternal
 {
@@ -342,7 +343,7 @@ void vesMapper::drawTriangles(const vesRenderState &renderState,
       numberOfIndicesToDraw = this->m_maximumTriangleIndicesPerDraw;
     }
 
-    unsigned int offset = 0;
+    uintptr_t offset = 0;
 
     // Send the primitive type information out
     renderState.m_material->bindRenderData(
